Move message class extensions to `:model`
