+++ java-tests/validating/src/test/kotlin/io/spine/test/options/goes/GoesOneWayITest.kt
@@ -29,6 +29,8 @@ package io.spine.test.options.goes
 import com.google.protobuf.Message
 import io.spine.test.options.set
 import io.spine.validate.ValidationException
+import io.spine.validation.getDescriptor
+import io.spine.validation.newBuilder
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
@@ -51,7 +53,7 @@ internal class GoesOneWayITest {
     @MethodSource("io.spine.test.options.goes.GoesOneWayTestEnv#onlyTargetFields")
     @ParameterizedTest(name = "throw if only the target `{1}` field is set")
     fun throwIfOnlyTargetFieldSet(message: Class<Message>, fieldName: String, fieldValue: Any) {
-        val descriptor = message.protoDescriptor()
+        val descriptor = message.getDescriptor()
         val field = descriptor.findFieldByName(fieldName)!!
         assertThrows<ValidationException> {
             message.newBuilder()
@@ -67,7 +69,7 @@ internal class GoesOneWayITest {
         fieldName: String,
         fieldValue: Any
     ) {
-        val descriptor = message.protoDescriptor()
+        val descriptor = message.getDescriptor()
         val companionField = descriptor.findFieldByName(fieldName)!!
         assertDoesNotThrow {
             message.newBuilder()
@@ -86,7 +88,7 @@ internal class GoesOneWayITest {
         fieldName: String,
         fieldValue: Any
     ) {
-        val descriptor = message.protoDescriptor()
+        val descriptor = message.getDescriptor()
         val field = descriptor.findFieldByName(fieldName)!!
         val companionField = descriptor.findFieldByName(companionName)!!
         assertDoesNotThrow {
